change format StaffHome
